fix bugs

+++ src/wishlists/wishlists.service.ts
@@ -1,34 +1,69 @@
-import { Injectable } from '@nestjs/common';
-import { CreateWishlistDto } from './dto/create-wishlist.dto';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { UpdateWishlistDto } from './dto/update-wishlist.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Wishlist } from './entities/wishlist.entity';
 import { Repository } from 'typeorm';
+import { User } from '../users/entities/user.entity';
+import { WishesService } from 'src/wishes/wishes.service';
 
 @Injectable()
 export class WishlistsService {
   constructor(
     @InjectRepository(Wishlist)
     private readonly wishlistRepository: Repository<Wishlist>,
+    private readonly wishesService: WishesService,
   ) {}
 
-  create(createWishlistDto: CreateWishlistDto): Promise<Wishlist> {
-    return this.wishlistRepository.save(createWishlistDto);
+  async create(user: User, createWishlistDto): Promise<Wishlist> {
+    for (let i = 0; i < createWishlistDto.itemsId.length; i++) {
+      const wish = await this.wishesService.findOne(
+        createWishlistDto.itemsId[i],
+      );
+      const wishlists = {
+        ...createWishlistDto,
+        owner: user,
+        items: [wish],
+      };
+      return await this.wishlistRepository.save(wishlists);
+    }
   }
 
-  findAll() {
-    return this.wishlistRepository.find();
+  async findAll() {
+    return await this.wishlistRepository.find();
   }
 
-  findOne(id: number): Promise<Wishlist> {
-    return this.wishlistRepository.findOneBy({ id });
+  async findOne(id: number): Promise<Wishlist> {
+    return await this.wishlistRepository.findOne({
+      where: {
+        id: id,
+      },
+      relations: ['owner', 'items'],
+    });
   }
 
-  update(id: number, updateWishlistDto: UpdateWishlistDto) {
-    return this.wishlistRepository.update({ id }, updateWishlistDto);
+  async update(
+    idUser: number,
+    id: number,
+    updateWishlistDto: UpdateWishlistDto,
+  ) {
+    const wishlists = await this.findOne(id);
+    if (wishlists.owner.id === idUser) {
+      return this.wishlistRepository.update({ id }, updateWishlistDto);
+    } else {
+      throw new UnauthorizedException(
+        'Вы не можете редактировать чужие списки подарков',
+      );
+    }
   }
 
-  remove(id: number) {
-    return this.wishlistRepository.delete({ id });
+  async remove(idUser: number, id: number) {
+    const wishlists = await this.findOne(id);
+    if (wishlists.owner.id === idUser) {
+      return await this.wishlistRepository.delete({ id });
+    } else {
+      throw new UnauthorizedException(
+        'Вы не можете удалять чужие списки подарков',
+      );
+    }
   }
 }
